pyproject.toml
[project]
name = "gemini-embedding-2-mcp-server"
version = "1.2.0"
description = "A multimodal local memory MCP for AI agents powered by Gemini Embedding 2"
authors = [
    { name = "Alaeddine Messadi" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "google-genai>=0.2.2",
    "chromadb>=0.4.24",
    "pymupdf>=1.23.0",
    "python-docx>=1.1.0"
]

[project.scripts]
gemini-embedding-2-mcp = "gemini_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gemini_mcp"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.12.0",
]
